Enable catching exceptions with Tyche
This commit demonstrates how to catch exceptions with Tyche.
The exceptions that cause a VM exit are configured using the VMX
exception bitmap, Tyche is able to modify the bitmap dynamically to add
more exceptions (e.g. #DE) by writing to the VMCS.

Currently Tyche needs to be explicitly called by the guest with an
hypercall 0x42, once called Tyche will setup the exception bitmap and
start catching exceptions (for now only #DE are supported).

A current limitation is that interrupts are not supported, even if the
interrupt ID correspond to one of the 32 exception IDs. This is likely
due to the current configuration of VTX.